+++ managed.go
@@ -1,0 +1,11 @@
+package pi
+
+import "github.com/joshp123/pi-golang/internal/sdk"
+
+func ClassifyManaged(result RunDetailedResult) ManagedSummary {
+	return sdk.ClassifyManaged(result)
+}
+
+func ClassifyRunError(err error) (BrokenCause, bool) {
+	return sdk.ClassifyRunError(err)
+}